modified:   datasets/Dataset_cls.py 	modified:   datasets/__pycache__/Dataset_cls.cpython-38.pyc 	modified:   main.py 	modified:   models/Transformer.py 	modified:   models/__pycache__/Transformer.cpython-38.pyc 	modified:   models/__pycache__/backbone.cpython-38.pyc 	modified:   models/__pycache__/detr.cpython-38.pyc 	modified:   models/backbone.py 	modified:   models/detr.py

+++ datasets/Dataset_cls.py
@@ -290,7 +290,7 @@ class LoadImagesAndLabels(Dataset):  # for training/testing
         }
         if nL:
             labels_out.update({
-                "labels": torch.as_tensor(labels[:, 0]),
+                "labels": torch.as_tensor(labels[:, 0]).long(),
                 "boxes": torch.as_tensor(labels[:, 1:5]),
                 "directions": torch.as_tensor(labels[:, 5:7])
             })
